refactor: 기본 예외 처리 메소드 클래스 구분
